LPS-96705 Move to DDM
